Day 4 part 2 speedup
Avoid bounds checking left by keeping right far enough out at all times,
and skip a bunch of comparisons in the process.

Note: Possible we miss a separator in the process, but if that was the
case, it was because the tag was malformed and the passport declaration
is invalid. If we miss the separator, it's going to be invalid anyway.